+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Unreleased
 - Added support for merging partial fills for same order and at same
   price to reduce number of generated entries
 - Added support for reporting regulatory fees
+  - Added logic for associating regulatory fees with the trades they
+    belong to
 - Made dividend and brokerage fee account names to use configurable
 - Bumped `apca` dependency to `0.19`